move fsEntryType from common/files to utils/files
src/utils/files/baseFS.ts and relativeFS.ts both imported
isFile/isDirectory from @common/files/fsEntryType, creating a
backward layering dependency: lower-level utils/ reaching into
higher-level common/. The helper itself has no VS Code coupling —
it only wraps the platform-agnostic FileType enum from
@platform/interfaces/filesystem — so it belongs in utils/files
alongside the other filesystem primitives.

Move src/common/files/fsEntryType.ts → src/utils/files/fsEntryType.ts
and update all 19 import sites. The common/files barrel re-exports the
symbols from the new location so callers that import from @common/files
(the index) continue to work without change.

Update CLAUDE.md and AGENTS.md to reference @utils/files/fsEntryType.
