+++ src/tools/memory/MemoryTool.ts
@@ -6,7 +6,7 @@ import { z } from 'zod';
 
 // Local imports
 import { tryUseRunContext } from '@agent/runtime/RunContext';
-import { isDirectory } from '@common/files/fsEntryType';
+import { isDirectory } from '@utils/files/fsEntryType';
 import { formatRelativeTime } from '@shared/utils/string';
 import { StorageFS } from '@utils/files';
 import { splitContentLines } from '@utils/text/stringUtils';
