+++ src/tools/memory/memoryFileSystem.ts
@@ -9,7 +9,7 @@ import * as path from 'path';
 
 import pMap from 'p-map';
 
-import { isDirectory, isSymlink } from '@common/files/fsEntryType';
+import { isDirectory, isSymlink } from '@utils/files/fsEntryType';
 import type { MemoryPreview, MemoryViewItem } from '@shared/schemas';
 import {
   MEMORY_STORAGE_ROOT,
