+++ src/utils/files/fsEntryType.ts
